pyproject.toml
[tool.poetry]
name = "test-project"
version = "0.1.0"
description = ""
authors = ["nickpashkevich <micolas9441@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
litestar = {extras = ["standard"], version = "^2.16.0"}
sqlalchemy = "^2.0.40"
uvicorn = "^0.34.2"
asyncpg = "^0.30.0"
black = "^25.1.0"
ruff = "^0.11.8"
advanced-alchemy = "^1.4.2"
pydantic = "^2.11.4"
loguru = "^0.7.3"
pydantic-settings = "^2.9.1"
pyhumps = "^3.8.0"
litestar-granian = "^0.12.3"
litestar-asyncpg = "^0.3.0"
msgspec = "^0.19.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"